requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
pyjwt
passlib
python-dotenv
pydantic
pydantic[email]
bcrypt
telethon